+++ src/vmware/azext_vmware/commands.py
@@ -54,7 +54,7 @@ def load_command_table(self, _):
         g.custom_show_command('show', 'datastore_show')
         g.custom_command('delete', 'datastore_delete')
 
-    with self.command_group('vmware datastore net-app-volume', vmware_sdk, client_factory=cf_vmware) as g:
+    with self.command_group('vmware datastore netapp-volume', vmware_sdk, client_factory=cf_vmware) as g:
         g.custom_command('create', 'datastore_netappvolume_create')
 
     with self.command_group('vmware datastore disk-pool-volume', vmware_sdk, client_factory=cf_vmware) as g: